+++ paparazzi/src/main/java/app/cash/paparazzi/DeviceConfig.kt
@@ -39,6 +39,7 @@ import com.android.resources.Keyboard
 import com.android.resources.KeyboardState
 import com.android.resources.Navigation
 import com.android.resources.NightMode
+import com.android.resources.NightMode.NOTNIGHT
 import com.android.resources.ScreenOrientation
 import com.android.resources.ScreenRatio
 import com.android.resources.ScreenSize
@@ -65,6 +66,7 @@ data class DeviceConfig(
   val xdpi: Int = 320,
   val ydpi: Int = 320,
   val orientation: ScreenOrientation = ScreenOrientation.PORTRAIT,
+  val nightMode: NightMode = NOTNIGHT,
   val density: Density = Density.XHIGH,
   val ratio: ScreenRatio = ScreenRatio.NOTLONG,
   val size: ScreenSize = ScreenSize.NORMAL,
@@ -93,7 +95,7 @@ data class DeviceConfig(
 
           updateScreenWidthAndHeight()
           uiModeQualifier = UiModeQualifier(UiMode.NORMAL)
-          nightModeQualifier = NightModeQualifier(NightMode.NOTNIGHT)
+          nightModeQualifier = NightModeQualifier(nightMode)
           countryCodeQualifier = CountryCodeQualifier()
           layoutDirectionQualifier = LayoutDirectionQualifier()
           networkCodeQualifier = NetworkCodeQualifier()
@@ -109,7 +111,7 @@ data class DeviceConfig(
 
   /**
    * Device specs per:
-   * https://android.googlesource.com/platform/tools/base/+/studio-master-dev/sdklib/src/main/java/com/android/sdklib/devices/nexus.xml
+   * https://android.googlesource.com/platform/tools/base/+/mirror-goog-studio-master-dev/sdklib/src/main/java/com/android/sdklib/devices/nexus.xml
    *
    * Release dates obtained from Wikipedia.
    */
@@ -353,6 +355,57 @@ data class DeviceConfig(
         softButtons = true,
         navigation = Navigation.NONAV,
         released = "May 7, 2019"
+    )
+
+    val PIXEL_4 = DeviceConfig(
+        screenHeight = 2280,
+        screenWidth = 1080,
+        xdpi = 444,
+        ydpi = 444,
+        orientation = ScreenOrientation.PORTRAIT,
+        density = Density.DPI_440,
+        ratio = ScreenRatio.LONG,
+        size = ScreenSize.NORMAL,
+        keyboard = Keyboard.NOKEY,
+        touchScreen = TouchScreen.FINGER,
+        keyboardState = KeyboardState.SOFT,
+        softButtons = true,
+        navigation = Navigation.NONAV,
+        released = "October 24, 2019"
+    )
+
+    val PIXEL_4_XL = DeviceConfig(
+        screenHeight = 3040,
+        screenWidth = 1440,
+        xdpi = 537,
+        ydpi = 537,
+        orientation = ScreenOrientation.PORTRAIT,
+        density = Density.DPI_560,
+        ratio = ScreenRatio.LONG,
+        size = ScreenSize.NORMAL,
+        keyboard = Keyboard.NOKEY,
+        touchScreen = TouchScreen.FINGER,
+        keyboardState = KeyboardState.SOFT,
+        softButtons = true,
+        navigation = Navigation.NONAV,
+        released = "October 24, 2019"
+    )
+
+    val PIXEL_4A = DeviceConfig(
+        screenHeight = 2340,
+        screenWidth = 1080,
+        xdpi = 442,
+        ydpi = 444,
+        orientation = ScreenOrientation.PORTRAIT,
+        density = Density.DPI_440,
+        ratio = ScreenRatio.LONG,
+        size = ScreenSize.NORMAL,
+        keyboard = Keyboard.NOKEY,
+        touchScreen = TouchScreen.FINGER,
+        keyboardState = KeyboardState.SOFT,
+        softButtons = true,
+        navigation = Navigation.NONAV,
+        released = "August 20, 2020"
     )
 
     private const val TAG_ATTR = "attr"