+++ CHANGELOG.md
@@ -1,8 +1,29 @@
 Change Log
 ==========
 
-## Version 0.1.0
+## Version 0.6.0
+_2020-10-02_
 
-_2019-06-28_
+As of this release, consumers must build on Java 11 environments.
 
-Initial release.
+### New
+* Point to a more recent version of layoutlib that runs on Android Q and builds with Java 11.
+* Refactor Paparazzi to better support non-Gradle builds
+* Added device configs for Pixel 4 series
+
+## Version 0.5.2
+_2020-09-17_
+
+### Fixed
+* [Gradle Plugin] Fixed record and verify tasks in multi-module projects.
+
+## Version 0.5.1
+_2020-09-17_
+
+### Fixed
+* [Gradle Plugin] Fixed race condition in record and verify tasks.
+
+## Version 0.5.0
+_2020-09-16_
+
+* Initial release.